Updated the script that creates the basic project structure to avoid creating existing elements and to support conditions.

+++ wordpress/Start-BasicProjectStructure.ps1
@@ -51,14 +51,32 @@ function Add-Item {
     $Item | Add-Member -MemberType NoteProperty -Name "path" -Value "$BasePath"
     $local:Path = [IO.Path]::Combine($Item.path, $Item.name)
 
-    # Create item
-    New-Item -Path $local:Path -ItemType $Item.type | Out-Null
-    
-    # Add default content if applies
-    if ($Item.type -eq "file" -AND (Get-Member -InputObject $Item -Name "default_content" -MemberType NoteProperty)) {
-        Get-DefaultContent $Item.default_content | Where-Object { $_ -ne $null } | Set-Content -Path $local:Path
+    # Check if condition is met
+    if (Get-Member -InputObject $Item -Name "condition" -MemberType NoteProperty) {
+        if ($Item.condition -eq "when-parent-not-empty") {
+            if (-Not (Test-Path "$BasePath\*")) {
+                $local:condition = $true
+            }
+            else {
+                $local:condition = $false
+            }
+        }
     }
-    
+    else {
+        $local:condition = $true # by default
+    }
+
+    # Only if the item DOES NOT exist and condition is met
+    if (-Not (Test-Path -Path $local:Path) -and $local:condition) {
+        # Create item
+        New-Item -Path $local:Path -ItemType $Item.type | Out-Null
+
+        # Add default content if applies
+        if ($Item.type -eq "file" -AND (Get-Member -InputObject $Item -Name "default_content" -MemberType NoteProperty)) {
+            Get-DefaultContent $Item.default_content | Where-Object { $_ -ne $null } | Set-Content -Path $local:Path
+        }        
+    }
+        
     # Iterate through children if any
     if (Get-Member -InputObject $Item -Name "children" -MemberType NoteProperty){
         $Item.children | ForEach-Object  {
